# Improve elixir implementation