Hello!
This PR improves the speed of transliteration by generating functions for each block in compile time and pattern match by them in runtime.
In addition, I added registration of all files in the module by `@external_resource` attribute and `__mix_recompile__?/0` for recompiling the module when changing or adding new blocks. It turned out that the perfomance problem was in reading and unzipping each block when processing incoming data.
At first I tried to make a `%{block_num => block_content}` map, but generating functions works much faster.
I have benchmarked code with text with 50/250/1000 letters and text with one letter repeated 350 times.
Below are inputs, benchmark code and results.

50 letters input:
```txt
Η θάλασσα είναι γαλήνια και ο ήλιος λάμπει.
```

250 letters input:
```txt
Η Ελλάδα είναι μια χώρα με πλούσια ιστορία και πολιτισμό. Οι αρχαίοι ναοί, η φιλοσοφία, και η δημοκρατία έχουν επηρεάσει όλο τον κόσμο. Τα νησιά της προσφέρουν μοναδικές εμπειρίες και εκπληκτικά τοπία για κάθε επισκέπτη.
```

1000 letters input:
```txt
Η Ελλάδα είναι γνωστή για την πλούσια ιστορία της, την πολιτιστική της κληρονομιά και τη φυσική της ομορφιά. Από την αρχαιότητα έως σήμερα, η χώρα διατηρεί τη μοναδικότητά της μέσα από τα μνημεία, την τέχνη, τη φιλοσοφία και την παράδοση. Οι αρχαίοι Έλληνες φιλόσοφοι, όπως ο Σωκράτης, ο Πλάτωνας και ο Αριστοτέλης, διαμόρφωσαν τη σκέψη της Δύσης και επηρέασαν βαθιά τις επιστήμες και τις ανθρωπιστικές σπουδές. Σήμερα, η Ελλάδα είναι δημοφιλής τουριστικός προορισμός, προσελκύοντας εκατομμύρια επισκέπτες κάθε χρόνο. Τα νησιά, όπως η Σαντορίνη, η Κρήτη και η Ρόδος, μαγεύουν με τα καταγάλανα νερά, την παραδοσιακή αρχιτεκτονική και τη φιλοξενία των κατοίκων. Επιπλέον, η ελληνική κουζίνα είναι ξακουστή σε όλο τον κόσμο, με πιάτα όπως ο μουσακάς, το σουβλάκι και η χωριάτικη σαλάτα. Η σύγχρονη Ελλάδα συνδυάζει το παρελθόν με το παρόν και συνεχίζει να εμπνέει με τον μοναδικό της χαρακτήρα.
```

repeated letter `α`:
```txt
ααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααααα
```

Benchmark code:
```elixir
text50 = File.read!("./50.txt")
text250 = File.read!("./250.txt")
text1000 = File.read!("./1000.txt")
repeat = File.read!("./repeat.txt")

bench = %{
  "persistent_term" => fn text -> AnyAscii.transliterate(text) end,
  "map" => fn text -> AnyAsciiMap.transliterate(text) end,
  "original" => fn text -> AnyAsciiPatternMatch.transliterate(text) end,
}

inputs = %{
  "50 symbols" => text50,
  "250 symbols" => text250,
  "1000 symbols" => text1000,
  "letter 'α' repeated 350 times" => repeat
}

Benchee.run(bench,
  inputs: inputs,
  parallel: 4,
  memory_time: 2,
  reduction_time: 2
)
```

Results:
```txt
Operating System: Linux
CPU Information: AMD Ryzen 7 PRO 4750U with Radeon Graphics
Number of Available Cores: 16
Available memory: 36.92 GB
Elixir 1.18.4
Erlang 27.3.4.1
JIT enabled: true

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 2 s
reduction time: 2 s
parallel: 4
inputs: 1000 symbols, 250 symbols, 50 symbols, letter 'α' repeated 350 times
Estimated total run time: 2 min 12 s

Benchmarking map with input 1000 symbols ...
Benchmarking map with input 250 symbols ...
Benchmarking map with input 50 symbols ...
Benchmarking map with input letter 'α' repeated 350 times ...
Benchmarking pattern_matching with input 1000 symbols ...
Benchmarking pattern_matching with input 250 symbols ...
Benchmarking pattern_matching with input 50 symbols ...
Benchmarking pattern_matching with input letter 'α' repeated 350 times ...
Benchmarking original with input 1000 symbols ...
Benchmarking original with input 250 symbols ...
Benchmarking original with input 50 symbols ...
Benchmarking original with input letter 'α' repeated 350 times ...
Calculating statistics...
Formatting results...

##### With input 1000 symbols #####
Name                       ips        average  deviation         median         99th %
pattern_matching       50.98 K       19.62 μs    ±41.03%       18.87 μs       28.06 μs
map                    31.41 K       31.84 μs    ±11.75%       30.81 μs       40.99 μs
original               14.78 K       67.64 μs     ±6.95%       67.11 μs       76.82 μs

Comparison: 
pattern_matching       50.98 K
map                    31.41 K - 1.62x slower +12.22 μs
original               14.78 K - 3.45x slower +48.03 μs

Memory usage statistics:

Name                Memory usage
pattern_matching           14 KB
map                        14 KB - 1.00x memory usage +0 KB
original                31.09 KB - 2.22x memory usage +17.09 KB

**All measurements for memory usage were the same**

Reduction count statistics:

Name             Reduction count
pattern_matching          5.16 K
map                       5.89 K - 1.14x reduction count +0.73 K
original                  5.93 K - 1.15x reduction count +0.77 K

**All measurements for reduction count were the same**

##### With input 250 symbols #####
Name                       ips        average  deviation         median         99th %
pattern_matching      178.11 K        5.61 μs   ±337.17%        5.43 μs        9.16 μs
map                   136.11 K        7.35 μs   ±262.19%        7.08 μs       13.81 μs
original               62.39 K       16.03 μs    ±50.66%       15.66 μs       23.24 μs

Comparison: 
pattern_matching      178.11 K
map                   136.11 K - 1.31x slower +1.73 μs
original               62.39 K - 2.85x slower +10.41 μs

Memory usage statistics:

Name                Memory usage
pattern_matching         3.52 KB
map                      3.52 KB - 1.00x memory usage +0 KB
original                 7.76 KB - 2.21x memory usage +4.24 KB

**All measurements for memory usage were the same**

Reduction count statistics:

Name             Reduction count
pattern_matching          1.24 K
map                       1.42 K - 1.15x reduction count +0.181 K
original                  1.45 K - 1.17x reduction count +0.21 K

**All measurements for reduction count were the same**

##### With input 50 symbols #####
Name                       ips        average  deviation         median         99th %
pattern_matching      821.68 K        1.22 μs  ±5975.07%        0.98 μs        1.97 μs
map                   638.30 K        1.57 μs  ±2057.68%        1.43 μs        2.35 μs
original              292.37 K        3.42 μs   ±909.03%        3.17 μs        4.54 μs

Comparison: 
pattern_matching      821.68 K
map                   638.30 K - 1.29x slower +0.35 μs
original              292.37 K - 2.81x slower +2.20 μs

Memory usage statistics:

Name                Memory usage
pattern_matching         1.42 KB
map                      1.42 KB - 1.00x memory usage +0 KB
original                 2.24 KB - 1.58x memory usage +0.82 KB

**All measurements for memory usage were the same**

Reduction count statistics:

Name             Reduction count
pattern_matching             242
map                          277 - 1.14x reduction count +35
original                     281 - 1.16x reduction count +39

**All measurements for reduction count were the same**

##### With input letter 'α' repeated 350 times #####
Name                       ips        average  deviation         median         99th %
pattern_matching      133.75 K        7.48 μs   ±161.42%        7.07 μs       14.92 μs
map                    90.00 K       11.11 μs    ±89.16%       10.53 μs       19.78 μs
original               34.87 K       28.68 μs    ±20.31%       28.19 μs       36.55 μs

Comparison: 
pattern_matching      133.75 K
map                    90.00 K - 1.49x slower +3.63 μs
original               34.87 K - 3.84x slower +21.20 μs

Memory usage statistics:

Name                Memory usage
pattern_matching         5.56 KB
map                      5.56 KB - 1.00x memory usage +0 KB
original                13.79 KB - 2.48x memory usage +8.23 KB

**All measurements for memory usage were the same**

Reduction count statistics:

Name             Reduction count
pattern_matching          2.26 K
map                       2.61 K - 1.16x reduction count +0.35 K
original                  2.66 K - 1.18x reduction count +0.40 K

**All measurements for reduction count were the same**
```