@hunterwb 2025-07-21
I am more concerned with file size and average memory usage. As Elixir packages are distributed as source code, this change adds 1.3 MB to the file size for users. The data is duplicated as the compressed files must still be present for the compilation and then they are also compiled into the beam file. The data in the beam file is in an expanded form as the files are ~250 KB compressed, ~500 KB uncompressed, or 1300 KB compiled.

The full data uses 4.5 MB of memory. This change loads all of the data into memory initially while otherwise it is loaded in blocks as needed. Most users will only need to load a couple of blocks.


I have made 3 relevant changes on branch `dev` (work in progress) since 0.3.3:

1. Use binary pattern matching instead of `String.to_charlist` for increased performance.

2. Increase block size from 256 to 4096, decreasing file size and decreasing the number of file reads and `persistent_term` additions. With 256 block size the total file size is 243 KB (2600 KB with filesystem overhead due to many small files), with 4096 block size it is 216 KB (315 KB with filesystem overhead). The number of blocks/files decreases from 640 to 48.

3. Reduce total memory usage from 4.5 MB to 2.7 MB by reusing duplicate strings in the same block, helped by the increased block size.